+++ packages/faceted-search/CHANGELOG.md
@@ -29,6 +29,7 @@ Types of changes
 ### Fixed
 
 - [fixed](https://github.com/Talend/ui/pull/2662): values open up at start
+- [fixed](https://github.com/Talend/ui/pull/2666): Put a scrollbar for enum badge with a lot of values
 
 ## [0.3.0]