+++ packages/faceted-search/CHANGELOG.md
@@ -31,6 +31,7 @@ Types of changes
 - [fixed](https://github.com/Talend/ui/pull/2665): sort alphabetically facets
 - [fixed](https://github.com/Talend/ui/pull/2662): values open up at start
 - [fixed](https://github.com/Talend/ui/pull/2666): Put a scrollbar for enum badge with a lot of values
+- [fixed](https://github.com/Talend/ui/pull/2676): Do not alter the filter input in checkboxes badge
 
 ## [0.3.0]