+++ packages/faceted-search/CHANGELOG.md
@@ -26,6 +26,10 @@ Types of changes
 
 ## [unreleased]
 
+### Changed
+
+- [Added](https://github.com/Talend/ui/pull/2673): Handle categories in badge definitions
+
 ### Added
 
 - [feat](https://github.com/Talend/ui/pull/2675): Empty screen text for search on badges list