+++ packages/faceted-search/CHANGELOG.md
@@ -26,6 +26,10 @@ Types of changes
 
 ## [unreleased]
 
+### Breaking change
+
+- ⤴️ @talend/ui: >= 4.32.0-0
+
 ### Changed
 
 - [Added](https://github.com/Talend/ui/pull/2673): Handle categories in badge definitions